Make addr_size (block manager's address sizes) a size_t instead of a uint32_t.
General review of uint32_t's, remove/cleanup where it's not necessary
to limit the type to 4B.

+++ src/block/block_addr.c
@@ -88,7 +88,7 @@ __wt_block_buffer_to_addr(WT_BLOCK *block,
  */
 int
 __wt_block_addr_valid(WT_SESSION_IMPL *session,
-    WT_BLOCK *block, const uint8_t *addr, uint32_t addr_size)
+    WT_BLOCK *block, const uint8_t *addr, size_t addr_size)
 {
 	off_t offset;
 	uint32_t cksum, size;
@@ -109,7 +109,7 @@ __wt_block_addr_valid(WT_SESSION_IMPL *session,
  */
 int
 __wt_block_addr_string(WT_SESSION_IMPL *session,
-    WT_BLOCK *block, WT_ITEM *buf, const uint8_t *addr, uint32_t addr_size)
+    WT_BLOCK *block, WT_ITEM *buf, const uint8_t *addr, size_t addr_size)
 {
 	off_t offset;
 	uint32_t cksum, size;
