+++ src/include/btmem.h
@@ -77,7 +77,7 @@ struct __wt_page_header {
  */
 struct __wt_addr {
 	uint8_t *addr;			/* Block-manager's cookie */
-	uint32_t size;			/* Block-manager's cookie length */
+	uint8_t  size;			/* Block-manager's cookie length */
 
 #define	WT_ADDR_INT	1		/* Internal page */
 #define	WT_ADDR_LEAF	2		/* Leaf page */
